Moved constraints to a new file constraints.go
